% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{AUC}
\alias{AUC}
\title{AUC}
\usage{
AUC(outcome, predict)
}
\arguments{
\item{outcome}{binary outcome (1: desired outcome; 0: otherwise)}

\item{predict}{prediction score}
}
\value{
a numeric value of empirical estimation of area under the ROC curves
}
\description{
Empirical AUC estimate
}
\details{
Function for AUC when input is X and Y.
}
\examples{
# no run
}
