% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{beta2theta}
\alias{beta2theta}
\title{beta2theta}
\usage{
beta2theta(beta)
}
\arguments{
\item{beta}{estimates of coefficient beta}
}
\value{
a numeric vector for theta (dimension-1)
}
\description{
Function to translate beta into theta, the n-sphere constrain
}
\details{
Function to translate beta into theta, the n-sphere constrain
}
\examples{
# no run
}
