% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{C.index}
\alias{C.index}
\title{C.index}
\usage{
C.index(yvar, score, censorvar, data)
}
\arguments{
\item{yvar}{column name for observed time}

\item{score}{column name for marker value}

\item{censorvar}{column name for censor (1 is event, 0 is censored)}

\item{data}{input data matrix}
}
\value{
a numeric value of empirical estimation of c-index
}
\description{
Empirical c-index estimate
}
\details{
Function for c-index when input is X and Y.
}
\examples{
# no run
}
