% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{cvfolds0}
\alias{cvfolds0}
\title{cvfolds0}
\usage{
cvfolds0(X, Y, idx, nfolds = 5)
}
\arguments{
\item{X}{marker matrix for non-responders}

\item{Y}{marker matrix for responders}

\item{idx}{m*n by 2 matrix for row index of marker matrix, first column is row index in X; second column is for Y}

\item{nfolds}{the cross-validation folds}
}
\value{
a vector containing CV fold index for each row in Z
}
\description{
internal function for generating CV fold index
}
\details{
Function for generate CV fold index
}
\examples{
# no run
}
