% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{grad.sub}
\alias{grad.sub}
\title{grad.sub}
\usage{
grad.sub(z, beta)
}
\arguments{
\item{z}{(m x n) x p data matrix as prepared for ROCSI}

\item{beta}{estimates of coefficient beta}
}
\value{
grad_square in the GCV
}
\description{
Internal function of grad_square in the GCV
}
\details{
Internal function of grad_square in the GCV
}
\examples{
# no run
}
