% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{gradsqr}
\alias{gradsqr}
\title{gradsqr}
\usage{
gradsqr(beta, Z0, index, w = 1)
}
\arguments{
\item{beta}{estimates of coefficient beta}

\item{Z0}{(m x n) x p Z matrix as prepared for ROCSI}

\item{index}{m*n by 2 matrix  for the subindex for the pair difference in Z}

\item{w}{a vector of weights Z (can be used for inverse probability weighting for missing data, default is 1)}
}
\value{
gradient square for the GCV.
}
\description{
Internal function for HIC calculation
}
\details{
Internal function for HIC calculation
}
\examples{
# no run
}
