% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{hessAUC}
\alias{hessAUC}
\title{hessAUC}
\usage{
hessAUC(beta, Z, w = 1)
}
\arguments{
\item{beta}{estimates of coefficient beta}

\item{Z}{(m x n) x p data matrix as prepared for ROCSI}

\item{w}{a vector of weights Z (can be used for inverse probability weighting for missing data, default is 1)}
}
\value{
Hessian matrix of AUC.
}
\description{
function for Hessian matrix of AUC
}
\details{
function for Hessian matrix of AUC
}
\examples{
# no run
}
