% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{hessAUC.sub}
\alias{hessAUC.sub}
\title{hessAUC.sub}
\usage{
hessAUC.sub(z, beta)
}
\arguments{
\item{z}{(m x n) x p data matrix as prepared for ROCSI}

\item{beta}{estimates of coefficient beta}
}
\value{
Hessian matrix components.
}
\description{
Internal function for hessAUC
}
\details{
Internal function for hessAUC
}
\examples{
# no run
}
