% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{HIC}
\alias{HIC}
\title{HIC}
\usage{
HIC(beta, Z, index, w = 1)
}
\arguments{
\item{beta}{estimates of coefficient beta}

\item{Z}{matrix prepared for ROCSI}

\item{index}{m*n by 2 matrix  for the subindex for the pair difference in Z}

\item{w}{a vector of weights Z (can be used for inverse probability weighting for missing data, default is 1)}
}
\value{
A numeric value with corresponding HIC
}
\description{
function for HIC calculation
}
\details{
Function for HIC calculation
}
\examples{
# no run
}
