% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{pair.diff}
\alias{pair.diff}
\title{pair.diff}
\usage{
pair.diff(X, Y, A)
}
\arguments{
\item{X}{marker matrix for non-responders}

\item{Y}{marker matrix for responders}

\item{A}{Treatment arm indicator (1 is treatment, 0 is control)}
}
\value{
A list of prepared data input for ROCSI
}
\description{
internal function for generating Z matrix (binary endpoint)
}
\details{
Function for generate Z matrix for binary endpoint
}
\examples{
# no run
}
