% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{pair.diff.surv}
\alias{pair.diff.surv}
\title{pair.diff.surv}
\usage{
pair.diff.surv(X, Y, A, C)
}
\arguments{
\item{X}{marker matrix}

\item{Y}{a vector for observed time}

\item{A}{a vector for Treatment arm indicator (1 is treatment, 0 is control)}

\item{C}{a vector for censor (1 is event, 0 is censored)}
}
\value{
A list of prepared data input for ROCSI
}
\description{
internal function for generating Z matrix (time-to-event endpoint)
}
\details{
Function for generate Z matrix for time-to-event endpoint
}
\examples{
# no run
}
