% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSI.R
\name{theta2beta}
\alias{theta2beta}
\title{theta2beta}
\usage{
theta2beta(theta)
}
\arguments{
\item{theta}{n-sphere coordination}
}
\value{
a numeric vector for beta (dimension+1)
}
\description{
Function to translate theta into beta
}
\details{
Function to translate beta into theta, the n-sphere constrain
}
\examples{
# no run
}
